Create polars.py
We are given the root 1+0j which is equivalent to the real number 1. This shows that all real number can be written as complex numbers in the form of their real part added to 0j. The two complex roots given are -0.5+((√3) /2)j and -0.5+(-(√3) /2)j. These three roots occur every 2/3π with a distance from the origin on 1. Logically, assumptions can be made regarding which positions on the coordinate plane will converge to which of the three roots. Any positions very close to a root will converge to that root quickly as this is how Newton's method works best. For points that are very close to the origin, their convergence will be chaotic. Their small z values ensure that their next approximation to the root will be much more inaccurate then their staring position as they will be sent to a position very far during the first iteration. Consequently they will not be good candidates to converge in few iterations because they are no longer approximate to any of the roots.